refactor: move shared helpers out of server
