+++ marimo/_internal/server/requests.py
@@ -1,13 +1,13 @@
 # Copyright 2026 Marimo. All rights reserved.
 """Internal API for server request types."""
 
-from marimo._server.models.export import (
+from marimo._schemas.export import (
     ExportAsHTMLRequest,
     ExportAsIPYNBRequest,
     ExportAsMarkdownRequest,
     ExportAsScriptRequest,
 )
-from marimo._server.models.models import InstantiateNotebookRequest
+from marimo._session.requests import InstantiateNotebookRequest
 
 __all__ = [
     "ExportAsHTMLRequest",
