+++ marimo/_server/api/endpoints/assets.py
@@ -32,13 +32,13 @@ from marimo._server.api.auth import TOKEN_QUERY_PARAM
 from marimo._server.api.deps import AppState
 from marimo._server.files.path_validator import PathValidator
 from marimo._server.router import APIRouter
-from marimo._server.templates.templates import (
+from marimo._session.model import SessionMode
+from marimo._templates import (
     home_page_template,
     inject_script,
     json_script,
     notebook_page_template,
 )
-from marimo._session.model import SessionMode
 from marimo._utils.async_path import AsyncPath
 from marimo._utils.paths import (
     MARIMO_DIR_NAME,
