+++ marimo/_server/models/models.py
@@ -33,7 +33,7 @@ from marimo._runtime.commands import (
     UpdateUserConfigCommand,
     ValidateSQLCommand,
 )
-from marimo._types.ids import CellId_t, UIElementId
+from marimo._types.ids import CellId_t
 
 
 class ListSecretKeysRequest(ListSecretKeysCommand, tag=False):
@@ -209,26 +209,6 @@ class CodeCompletionRequest(CodeCompletionCommand, tag=False):
             document=self.document,
             cell_id=self.cell_id,
         )
-
-
-class UpdateUIElementValuesRequest(msgspec.Struct, rename="camel"):
-    object_ids: list[UIElementId]
-    values: list[Any]
-
-    # Validate same length
-    def __post_init__(self) -> None:
-        assert len(self.object_ids) == len(self.values), (
-            "Mismatched object_ids and values"
-        )
-
-
-class InstantiateNotebookRequest(UpdateUIElementValuesRequest):
-    auto_run: bool = True
-    # Optional: cell codes to use instead of the codes from the file.
-    # This is used when the frontend has local edits that should be
-    # used instead of the file codes (e.g., pre-connect editing).
-    # Maps cell_id -> code.
-    codes: dict[CellId_t, str] | None = None
 
 
 class BaseResponse(msgspec.Struct, rename="camel"):
